+++ app/controllers/reviews_controller.rb
@@ -4,6 +4,8 @@ class ReviewsController < ApplicationController
 
   def create
     @book = Book.find(params[:book_id])
+    review_params = params.require(:review).permit(:comment, :star, :user_id)
+    review_params[:user_id] = current_user.id
     if @book.reviews.create(review_params)
       redirect_to book_path(@book)
     else
@@ -16,6 +18,7 @@ class ReviewsController < ApplicationController
   end
 
   def update
+    authorize @review, policy_class: ReviewPolicy
     if @review.update(review_params)
       redirect_to book_path(@review.book)
     else
@@ -24,6 +27,7 @@ class ReviewsController < ApplicationController
   end
 
   def destroy
+    authorize @review, policy_class: ReviewPolicy
     if @review.destroy
       redirect_to @review.book
     else
